requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
numpy==1.22.1
pandas==1.4.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
